requirements.txt
aiosqlite==0.17.0
discord.py==2.0.1
discord-sentry-reporting==2.0.4
python-dotenv==0.21.0
sentry-sdk==1.9.10
tzlocal==4.2
SQLAlchemy==1.4.41
openpyxl
requests
aiohttp
Pillow